[enhanc] merge algorithms.py code to gradient.py
